+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## unreleased
 * Increase valid Discover card length to 19 digits
 * Require `braintree_ios` v5.4.1 or higher
+* iOS 15 Support
+  * Update `scrollEdgeAppearance` based on iOS 15 changes
 
 ## 9.0.2 (2021-05-26)
 * Require `braintree_ios` v5.3.2 or higher